new lock file
